custom_components/ha_cloud_music/requirements.txt
aiohttp>=3.8.0
pycryptodome>=3.15.0 